10:27 Co-authored-by: Divyansh Sharma <div360@users.noreply.github.com>
