+++ react-projects/samvadhini/package.json
@@ -9,7 +9,12 @@
     "axios": "^0.27.2",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "react-tsparticles": "^2.1.3",
+    "react-typist": "^2.0.5",
+    "tsparticles": "^2.1.3",
+    "typed.js": "^2.0.12",
     "web-vitals": "^2.1.0"
   },
   "scripts": {
