+++ react-projects/samvadhini/src/App.js
@@ -3,8 +3,12 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Home from './pages/Home/Home';
 import Model from './pages/Model/Model';
 import Thanks from './pages/TU/TU';
+
 const App =()=>{
+
+
   return(
+    <div className="App">
     <BrowserRouter>
     <Routes>
       <Route exact path ='/' element={<Home/>}/>
@@ -12,6 +16,7 @@ const App =()=>{
       <Route exact path= '/thankyou' element={<Thanks />}/>
     </Routes>
     </BrowserRouter>
+    </div>
   )
 }
 export default App
