+++ react-projects/samvadhini/src/pages/Model/model.css
@@ -1,0 +1,11 @@
+.container{
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    margin-right: -50%;
+    transform: translate(-50%, -50%);
+    background:white;
+    width: 80vw;
+    height: 80vh;
+    border-radius: 20px;
+}
